Pseudo parser now sends all the modifiers as expected. Though we get a query too complex error
