+++ src/Sidekick.Apis.Poe/Parser/ItemParser.cs
@@ -7,8 +7,8 @@ using Sidekick.Apis.Poe.Parser.Metadata;
 using Sidekick.Apis.Poe.Parser.Modifiers;
 using Sidekick.Apis.Poe.Parser.Patterns;
 using Sidekick.Apis.Poe.Parser.Properties;
+using Sidekick.Apis.Poe.Parser.Pseudo;
 using Sidekick.Apis.Poe.Parser.Sockets;
-using Sidekick.Apis.Poe.Pseudo;
 using Sidekick.Common.Exceptions;
 using Sidekick.Common.Game.Items;
 
@@ -19,7 +19,7 @@ namespace Sidekick.Apis.Poe.Parser
         ILogger<ItemParser> logger,
         IMetadataParser metadataProvider,
         IModifierParser modifierParser,
-        IPseudoModifierProvider pseudoModifierProvider,
+        IPseudoParser pseudoParser,
         IParserPatterns patterns,
         ClusterJewelParser clusterJewelParser,
         IInvariantMetadataProvider invariantMetadataProvider,
@@ -71,7 +71,7 @@ namespace Sidekick.Apis.Poe.Parser
                 var sockets = socketParser.Parse(parsingItem);
                 var modifierLines = ParseModifiers(parsingItem);
                 var properties = propertyParser.Parse(parsingItem, modifierLines);
-                var pseudoModifiers = pseudoModifierProvider.Parse(modifierLines);
+                var pseudoModifiers = pseudoParser.Parse(modifierLines);
                 var item = new Item(metadata: metadata,
                                     invariant: invariant,
                                     header: header,
